pyproject.toml
[project]
name = "phi_cloud_action"
version = "1.5.4.1b"
description = "使用Python实现的Phigros云端数据操作"
authors = [
    {name = "wms26", email = "2652015696@qq.com"},
]
dependencies = [ 
    "pycryptodome>=3.21.0", 
    "colorlog>=6.9.0", 
    "fastapi>=0.115.8", 
    "uvicorn>=0.33.0", 
    "pyyaml>=6.0.2", 
    "tqdm>=4.67.1", 
    "beautifulsoup4>=4.13.3", 
    "setuptools>=75.3.0", 
    "PyGithub>=2.6.0", 
    "pyqrcode>=1.2.1", 
    "httpx<=0.27", 
    "jinja2>=3.1.5",
    "packaging>=24.2",
    "pydantic>=2.10.6",
    "playwright>=1.50.0"
    ]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "GPL-3.0"}

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
phi_cloud_action = [
    "data/example/**/*",
    "data/info/**/*",
    "data/*.yml",
    "data/web/resources/resources",
    ]

[tool.setuptools.packages]
find = {where = ["."], exclude = ["tests*"]}

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pdm.scripts]
test = "pytest -s"  # 测试脚本
safe_test = "pytest" # 安全测试
dev = {call = "phi_cloud_action.webapi:main", env = {DEV = "true"} } # dev启动
start = {call = "phi_cloud_action.webapi:main"} # 正常启动
uninstall = "pdm remove" # 卸载
install = "pdm add" # 安装
build = "pdm build" # 打包